Move classes to it's own files.
